Doctests and grouping files
