--- Python/area_of_circle.py
@@ -1,8 +1,0 @@
-from math import pi
-from math import pow
-
-def area_circle(radius):
-    """
-    Calculates the area of a circle given the radius
-    """
-    return pi * pow(radius,2)
