+++ README.md
@@ -1,4 +1,5 @@
 # simple-maths
+
 Simple math functions in Python. (As of now, at least).
 
 This project will be dedicated to create (mostly) simple (and hopefully) fast mathematical functions.
