+++ src/combinatorics.py
@@ -1,3 +1,4 @@
+
 def combination(n,r):
     """
     Function to calculate combination of 'n' objects, selected 'r' at a time.
@@ -14,3 +15,18 @@ def combination(n,r):
         denominator *= i
         N -= 1
     return int(numerator / denominator)
+
+
+def permutation(n, r):
+    """
+    Function to calculate permuation of 'n' objects, selected 'r' at a time.
+    """
+    result = 1
+    for i in range(n, n-r, -1):
+        result *= i
+    return result
+
+
+if __name__ == '__main__':
+    import doctest
+    doctest.testmod()
